requirements.txt
tensorflow==2.12.0
matplotlib
scikit-learn
numpy
pillow
flask
gradio
